move math package one level up
