+++ src/main/java/com/fundynamic/d2tm/game/behaviors/Selectable.java
@@ -1,7 +1,11 @@
 package com.fundynamic.d2tm.game.behaviors;
 
-/**
- * Created by shendriks on 20/02/15.
- */
 public interface Selectable {
+
+    void select();
+
+    void deselect();
+
+    boolean isSelected();
+
 }
