+++ src/main/java/com/fundynamic/d2tm/game/entities/units/UnitsRepository.java
@@ -2,7 +2,7 @@ package com.fundynamic.d2tm.game.entities.units;
 
 
 import com.fundynamic.d2tm.game.map.Map;
-import com.fundynamic.d2tm.game.math.Vector2D;
+import com.fundynamic.d2tm.math.Vector2D;
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
