+++ src/main/java/com/fundynamic/d2tm/game/map/Map.java
@@ -1,7 +1,7 @@
 package com.fundynamic.d2tm.game.map;
 
 import com.fundynamic.d2tm.game.map.renderer.TerrainFacingDeterminer;
-import com.fundynamic.d2tm.game.math.Vector2D;
+import com.fundynamic.d2tm.math.Vector2D;
 import com.fundynamic.d2tm.game.entities.structures.Structure;
 import com.fundynamic.d2tm.game.entities.structures.StructuresRepository;
 import com.fundynamic.d2tm.game.terrain.Terrain;
