+++ src/main/java/com/fundynamic/d2tm/game/map/renderer/MapEntityViewportRenderer.java
@@ -4,7 +4,7 @@ package com.fundynamic.d2tm.game.map.renderer;
 import com.fundynamic.d2tm.game.map.Cell;
 import com.fundynamic.d2tm.game.map.Map;
 import com.fundynamic.d2tm.game.map.MapEntity;
-import com.fundynamic.d2tm.game.math.Vector2D;
+import com.fundynamic.d2tm.math.Vector2D;
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
