+++ src/main/java/com/fundynamic/d2tm/game/map/renderer/ViewportRenderer.java
@@ -1,6 +1,6 @@
 package com.fundynamic.d2tm.game.map.renderer;
 
-import com.fundynamic.d2tm.game.math.Vector2D;
+import com.fundynamic.d2tm.math.Vector2D;
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
