+++ src/test/java/com/fundynamic/d2tm/game/map/PerimeterTest.java
@@ -1,6 +1,6 @@
 package com.fundynamic.d2tm.game.map;
 
-import com.fundynamic.d2tm.game.math.Vector2D;
+import com.fundynamic.d2tm.math.Vector2D;
 import junit.framework.Assert;
 import org.junit.Test;
