+++ src/test/java/com/fundynamic/d2tm/game/map/renderer/MapCellViewportRendererTest.java
@@ -3,7 +3,7 @@ package com.fundynamic.d2tm.game.map.renderer;
 import com.fundynamic.d2tm.game.map.CellFactory;
 import com.fundynamic.d2tm.game.map.Map;
 import com.fundynamic.d2tm.game.map.MapCell;
-import com.fundynamic.d2tm.game.math.Vector2D;
+import com.fundynamic.d2tm.math.Vector2D;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
